lib: move code around, do renamings
For coherence, we move some code around. Out of separate utility files
and into the only files that import them.

We rename some functions for consistency.

We expose all the aura-specific utilities under objects named after
their respective auras.

We lightly improve doccomment consistency.
