+++ src/index.ts
@@ -4,5 +4,7 @@ export { parse, tryParse, valid, slav, slaw, nuck } from './parse';
 export { render, scot, rend } from './render';  //TODO  expose encodeString() ?
 
 //  atom utils
-export { daToUnix, unixToDa } from './da';
-export { cite, deSig, preSig } from './p';  //TODO  remove deSig, preSig
+import { toUnix, fromUnix } from './da';
+export const da = { toUnix, fromUnix };
+import { cite, sein, clan, kind, rankToSize, sizeToRank } from './p';
+export const p = { cite, sein, clan, kind, rankToSize, sizeToRank };
